+++ src/nvim/ui.c
@@ -639,8 +639,8 @@ void ui_grid_resize(handle_T grid_handle, int width, int height, Error *error)
     }
   } else {
     // non-positive indicates no request
-    wp->w_height_request = (int)MAX(height, 0);
-    wp->w_width_request = (int)MAX(width, 0);
+    wp->w_height_request = MAX(height, 0);
+    wp->w_width_request = MAX(width, 0);
     win_set_inner_size(wp);
   }
 }
